moved logic out of app component to keep it separate and make component readable
